Implement exchangeable Qt-OpenGL backend.
The default backend is based on QOpenGLWidget, which was also used
before. Since this has certain limitations such as it does not support
stereo rendering, another backend was implemented, based on
QOpenGLWindow.
As it turns out, this is also much more performant in some cases.
However, some testing is required before we can consider this as
default backend.