Add option to enqueue invalidation in AsyncComputeProcessor
The user can select between the current (default) behaviour (abort on
any input data change) and lazy recomputation of the result after the
current computation has finished, but only if this is possible for the
data that changed. In particular, any changes to data that requires
invalidation observers (such as volumes) still cause an abort. On the
other hand, data that is copied for thread safe access (such as
geometries) is safe to use after the input port's data has changed and
only enqueues an update.

This is especially useful for RandomWalker/OctreeWalker users, which can
just keep changing the labels while in the background a new output
segmentation is computed seamlessly.