Improve usability of OctreeCreator
Previously, it was not allowed to create an octree from input volumes
that already have an octree representation. This (presumably) was an
attempt to hinder the creation of multiple octrees at the same time,
which is not handled too well by the Disk Brickpoolmanager. The Mmap
version handles that just fine, though, so we don't want to restrict its
usage.

Additionally, checking if output is up to date/incompatible has been
greatly simplified for both single and multi volume inputs.

The "Clear Octree" button now *actually* clears the octree and doesn't
just delete its files from under its feet.