Allow rendering of octrees deeper than 8 levels
At least 9 levels are supported. Theoretically, even deeper octrees (for
example, for highly non-cubic volumes) are also supported, although the
octree construction still allocates a lot of memory (assuming a
power-of-two cubic base) for them.

We buy the deeper depth by reducing the avg value precision for 3/4
channel volumes by a bit (literally), but I think it's still worth it.